+++ install.ps1
@@ -1,4 +1,4 @@
-$version="0.60.3"
+$version="0.61.0"
 
 $fzf_base=Split-Path -Parent $MyInvocation.MyCommand.Definition