+++ CHANGELOG.md
@@ -10,6 +10,7 @@ CHANGELOG
   ```
 - Added `change-ghost` and `transform-ghost` actions for dynamically changing the ghost text
 - Added `change-pointer` and `transform-pointer` actions for dynamically changing the pointer sign
+- Added `r` flag for placeholder expression (raw mode) for unquoted output
 - Bug fixes and improvements
 
 0.60.3